--- specs/resources/fake_project_with_nyc_report_file/nyc-coverage/lcov.info
@@ -1,314 +1,0 @@
-TN:
-SF:E:\work\sonar-scanner-npm\src\sonarqube-scanner-executable.js
-FN:22,prepareExecEnvironment
-FN:56,getSonarQubeScannerExecutable
-FN:103,(anonymous_2)
-FN:118,getLocalSonarQubeScannerExecutable
-FN:142,findTargetOS
-FN:158,isWindows
-FN:162,isMac
-FN:166,isLinux
-FNF:8
-FNH:1
-FNDA:2,prepareExecEnvironment
-FNDA:0,getSonarQubeScannerExecutable
-FNDA:0,(anonymous_2)
-FNDA:0,getLocalSonarQubeScannerExecutable
-FNDA:0,findTargetOS
-FNDA:0,isWindows
-FNDA:0,isMac
-FNDA:0,isLinux
-DA:1,1
-DA:2,1
-DA:3,1
-DA:4,1
-DA:5,1
-DA:6,1
-DA:7,1
-DA:8,1
-DA:9,1
-DA:10,1
-DA:11,1
-DA:13,1
-DA:14,1
-DA:15,1
-DA:24,2
-DA:25,2
-DA:26,1
-DA:28,2
-DA:35,2
-DA:36,2
-DA:41,2
-DA:50,2
-DA:57,0
-DA:58,0
-DA:59,0
-DA:60,0
-DA:61,0
-DA:62,0
-DA:64,0
-DA:70,0
-DA:71,0
-DA:72,0
-DA:73,0
-DA:75,0
-DA:76,0
-DA:78,0
-DA:80,0
-DA:81,0
-DA:82,0
-DA:87,0
-DA:88,0
-DA:89,0
-DA:90,0
-DA:91,0
-DA:96,0
-DA:97,0
-DA:98,0
-DA:99,0
-DA:104,0
-DA:105,0
-DA:106,0
-DA:107,0
-DA:108,0
-DA:109,0
-DA:111,0
-DA:119,0
-DA:120,0
-DA:121,0
-DA:125,0
-DA:126,0
-DA:127,0
-DA:130,0
-DA:131,0
-DA:132,0
-DA:135,0
-DA:143,0
-DA:144,0
-DA:146,0
-DA:147,0
-DA:149,0
-DA:150,0
-DA:152,0
-DA:159,0
-DA:163,0
-DA:167,0
-LF:75
-LH:22
-BRDA:25,0,0,1
-BRDA:25,0,1,1
-BRDA:61,1,0,0
-BRDA:61,1,1,0
-BRDA:80,2,0,0
-BRDA:80,2,1,0
-BRDA:90,3,0,0
-BRDA:90,3,1,0
-BRDA:104,4,0,0
-BRDA:104,4,1,0
-BRDA:120,5,0,0
-BRDA:120,5,1,0
-BRDA:143,6,0,0
-BRDA:143,6,1,0
-BRDA:146,7,0,0
-BRDA:146,7,1,0
-BRDA:149,8,0,0
-BRDA:149,8,1,0
-BRF:18
-BRH:2
-end_of_record
-TN:
-SF:E:\work\sonar-scanner-npm\src\sonarqube-scanner-params.js
-FN:14,defineSonarQubeScannerParams
-FN:64,extractInfoFromPackageFile
-FNF:2
-FNH:2
-FNDA:10,defineSonarQubeScannerParams
-FNDA:2,extractInfoFromPackageFile
-DA:1,1
-DA:2,1
-DA:3,1
-DA:4,1
-DA:5,1
-DA:6,1
-DA:8,1
-DA:16,10
-DA:17,10
-DA:18,10
-DA:19,10
-DA:23,9
-DA:33,9
-DA:34,9
-DA:35,9
-DA:37,2
-DA:40,7
-DA:41,7
-DA:46,10
-DA:47,4
-DA:51,10
-DA:52,2
-DA:54,10
-DA:55,2
-DA:57,10
-DA:58,1
-DA:61,10
-DA:65,2
-DA:66,2
-DA:67,2
-DA:68,2
-DA:69,2
-DA:70,2
-DA:71,2
-DA:72,1
-DA:74,2
-DA:75,1
-DA:77,2
-DA:78,1
-DA:80,2
-DA:81,1
-DA:83,2
-DA:84,2
-DA:86,0
-DA:88,0
-DA:90,0
-DA:92,0
-DA:94,0
-DA:98,0
-DA:103,0
-LF:50
-LH:43
-BRDA:46,0,0,4
-BRDA:46,0,1,6
-BRDA:51,1,0,2
-BRDA:51,1,1,8
-BRDA:54,2,0,2
-BRDA:54,2,1,8
-BRDA:57,3,0,1
-BRDA:57,3,1,9
-BRDA:67,4,0,2
-BRDA:67,4,1,0
-BRDA:71,5,0,1
-BRDA:71,5,1,1
-BRDA:74,6,0,1
-BRDA:74,6,1,1
-BRDA:77,7,0,1
-BRDA:77,7,1,1
-BRDA:77,8,0,2
-BRDA:77,8,1,1
-BRDA:80,9,0,1
-BRDA:80,9,1,1
-BRDA:80,10,0,2
-BRDA:80,10,1,1
-BRDA:83,11,0,2
-BRDA:83,11,1,0
-BRDA:86,12,0,0
-BRDA:86,12,1,0
-BRDA:90,13,0,0
-BRDA:90,13,1,0
-BRDA:90,14,0,0
-BRDA:90,14,1,0
-BRDA:94,15,0,0
-BRDA:94,15,1,0
-BRDA:94,16,0,0
-BRDA:94,16,1,0
-BRDA:98,17,0,0
-BRDA:98,17,1,0
-BRDA:98,17,2,0
-BRDA:98,17,3,0
-BRDA:98,17,4,0
-BRF:39
-BRH:22
-end_of_record
-TN:
-SF:E:\work\sonar-scanner-npm\specs\sonarqube-scanner-executable-test.js
-FN:5,(anonymous_0)
-FN:7,(anonymous_1)
-FN:25,(anonymous_2)
-FN:35,(anonymous_3)
-FN:60,(anonymous_4)
-FN:72,pathForProject
-FNF:6
-FNH:6
-FNDA:1,(anonymous_0)
-FNDA:1,(anonymous_1)
-FNDA:1,(anonymous_2)
-FNDA:1,(anonymous_3)
-FNDA:1,(anonymous_4)
-FNDA:2,pathForProject
-DA:1,1
-DA:2,1
-DA:3,1
-DA:5,1
-DA:7,1
-DA:8,1
-DA:23,1
-DA:26,1
-DA:30,1
-DA:35,1
-DA:36,1
-DA:53,1
-DA:61,1
-DA:65,1
-DA:73,2
-LF:15
-LH:15
-BRF:0
-BRH:0
-end_of_record
-TN:
-SF:E:\work\sonar-scanner-npm\specs\sonarqube-scanner-params-test.js
-FN:5,(anonymous_0)
-FN:7,(anonymous_1)
-FN:21,(anonymous_2)
-FN:28,(anonymous_3)
-FN:47,(anonymous_4)
-FN:65,(anonymous_5)
-FN:79,(anonymous_6)
-FN:96,(anonymous_7)
-FN:115,(anonymous_8)
-FN:136,pathForProject
-FNF:10
-FNH:10
-FNDA:1,(anonymous_0)
-FNDA:1,(anonymous_1)
-FNDA:1,(anonymous_2)
-FNDA:1,(anonymous_3)
-FNDA:1,(anonymous_4)
-FNDA:1,(anonymous_5)
-FNDA:1,(anonymous_6)
-FNDA:1,(anonymous_7)
-FNDA:1,(anonymous_8)
-FNDA:8,pathForProject
-DA:1,1
-DA:2,1
-DA:3,1
-DA:5,1
-DA:7,1
-DA:8,1
-DA:16,1
-DA:21,1
-DA:22,1
-DA:23,1
-DA:28,1
-DA:29,1
-DA:39,1
-DA:47,1
-DA:48,1
-DA:57,1
-DA:65,1
-DA:66,1
-DA:74,1
-DA:79,1
-DA:80,1
-DA:91,1
-DA:96,1
-DA:97,1
-DA:107,1
-DA:115,1
-DA:116,1
-DA:126,1
-DA:137,8
-LF:29
-LH:29
-BRF:0
-BRH:0
-end_of_record
