+++ test/integration/resources/fake_project_for_integration/package.json
