+++ test/unit/index.test.js
@@ -1,10 +1,10 @@
 const assert = require('assert');
-const index = require('../src/index');
+const index = require('../../src/index');
 
 describe('fromParam', function () {
   it('should provide the correct identity', function () {
     assert.deepEqual(index.fromParam(), [
-      '--from=ScannerNpm/' + require('../package.json').version,
+      '--from=ScannerNpm/' + require('../../package.json').version,
     ]);
   });
 });
