+++ test/unit/resources/fake_project_with_basic_package_file/package.json
