+++ test/unit/resources/fake_project_with_complete_package_file/package.json
