+++ test/unit/resources/fake_project_with_jest_report_file/package.json
