+++ test/unit/resources/fake_project_with_no_package_file/index.js
