+++ test/unit/resources/fake_project_with_nyc_report_file/nyc-coverage/lcov.info
