+++ test/unit/resources/fake_project_with_nyc_report_file/package.json
