+++ test/unit/resources/fake_project_with_scoped_package_name/package.json
