+++ test/unit/resources/fake_project_with_sonar_properties_file/sonar-project.properties
